requirements.txt
annotated-types==0.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
python-dotenv==1.0.1
requests==2.32.2
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
